feat: entity setting
