+++ src/main/java/roomescape/reservation/model/entity/Reservation.java
@@ -1,22 +1,45 @@
-package roomescape.domain.reservation.model.entity;
+package roomescape.reservation.model.entity;
 
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.Objects;
+import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
-import roomescape.domain.reservation.model.dto.ReservationDetails;
-import roomescape.domain.reservation.model.exception.ReservationException.InvalidReservationTimeException;
+import lombok.NoArgsConstructor;
+import roomescape.reservation.model.dto.ReservationDetails;
+import roomescape.reservation.model.exception.ReservationException.InvalidReservationTimeException;
 
 @Getter
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class Reservation {
 
-    private final Long id;
-    private final String name;
-    private final LocalDate date;
-    private final ReservationTime time;
-    private final ReservationTheme theme;
-    private final Long memberId;
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false)
+    private String name;
+
+    @Column(nullable = false)
+    private LocalDate date;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    private ReservationTime time;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    private ReservationTheme theme;
+
+    private Long memberId;
 
     @Builder
     public Reservation(Long id, String name, LocalDate date, ReservationTime time, ReservationTheme theme,
